Attempt using number-format
The NumberFormat text input doesn't really seem
set up for use with react-form-hook. I would have to
use the onValueChange handler of NumberFormat
to manually update the state field if I want to store
the unformatted value, or I have to remove the $s and ,s
on save and recheck for validity. I might be missing something